Change Logic Instructions values to be more meaningful
On the "Logic Instructions" map, the values used in the example
did not cause the result to ever become zero, hence the student
could not observe the Z bit changing.  In addition, the text
indicated that the W would contain 0xf8  before ANDing with the
STATUS register.  This was not the case.